refactor: adopt module based folder structure with path aliases
