+++ pages/index.tsx
@@ -1,7 +1,5 @@
-import Head from 'next/head'
-import Image from 'next/image'
-import styles from '../styles/Home.module.css'
-import Landing from './landing'
+import '@/styles/Home.module.css'
+import Landing from '@/components/elements/Landing'
 
 export default function Home() {
 	return <Landing />
