+++ src/common/components/elements/Link.tsx
@@ -1,7 +1,7 @@
 
 // @ts-nocheck 
 import React from 'react';
-import sv from '../../constants/Styles'
+import sv from '@/constants/Styles'
 import Link from 'next/link';
 import styled from 'styled-components';
