+++ src/common/components/elements/LoadingTest.tsx
@@ -1,6 +1,6 @@
-import sv from '../constants/Styles'
+import sv from '@/constants/Styles'
 import styled from 'styled-components';
-import FullPageLoading from '../components/core/FullPageLoading';
+import FullPageLoading from './FullPageLoading';
 // @ts-ignore
 import { rgba } from 'polished';
