+++ src/modules/storefront/client.ts
@@ -1,6 +1,5 @@
 import React from "react";
-import { StorefrontTheme, Storefront } from "../../../types";
-
+import { StorefrontTheme, Storefront } from "./types";
  
 export async function checkStorefrontAvailability(subdomain: string, dispatch: React.Dispatch<any>) {
   const requestOptions = {
@@ -21,7 +20,7 @@ export async function checkStorefrontAvailability(subdomain: string, dispatch: R
           desiredStorefrontSubdomain: subdomain
         },
       })
-      
+
       return
     }
 
@@ -96,6 +95,4 @@ export async function saveTheme(
    } catch (error) {
      dispatch({ type: 'THEME_SAVE_ERROR', payload: { error }})
    }
-
-
 }
